# backend/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.Basket:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      expiration_date:
        type: string
      id:
        type: integer
      name:
        type: string
      original_price:
        type: number
      price:
        type: number
      quantity:
        type: integer
      restaurant:
        $ref: '#/definitions/models.Restaurant'
      restaurant_id:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
    required:
    - name
    - original_price
    - price
    - quantity
    - type
    type: object
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Merchant:
    properties:
      business_name:
        description: Nom de l'entreprise
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email_pro:
        description: Email valide requis
        type: string
      id:
        type: integer
      phone_number:
        description: Numéro de téléphone (optionnel, max 15 caractères)
        type: string
      siret:
        description: Numéro SIRET
        type: string
      updatedAt:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relation vers User (clé étrangère avec cascade)
      user_id:
        description: Relation 1 à 1 vers User
        type: integer
    required:
    - business_name
    - email_pro
    - siret
    type: object
...
  models.Restaurant:
    properties:
      address:
        description: Adresse complète
        type: string
      category:
        allOf:
        - $ref: '#/definitions/models.Category'
        description: Relation avec Category (clé étrangère)
      category_id:
        description: ID de la catégorie (clé étrangère)
        type: integer
      city:
        description: Ville
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      merchant:
        allOf:
        - $ref: '#/definitions/models.Merchant'
        description: Relation avec le commerçant
      merchant_id:
        description: ID du commerçant (clé étrangère)
        type: integer
      name:
        description: Nom du restaurant (obligatoire)
        type: string
      phone_number:
        description: Numéro de téléphone (optionnel, max 15 caractères)
        type: string
      postal_code:
        description: Code postal (limité à 10 caractères pour compatibilité internationale)
        type: string
      rating:
        description: Note moyenne (sur 5)
        type: number
      siren:
        description: SIREN (exactement 9 chiffres, unique)
        type: string
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: Validation d'email
        type: string
      full_name:
        description: Nom complet requis
        type: string
      id:
        type: integer
      is_admin:
        description: Est-ce un administrateur ?
        type: boolean
      password_hash:
        description: Hash du mot de passe
        type: string
      phone:
        description: Téléphone requis
        type: string
      updatedAt:
        type: string
    required:
    - email
    - full_name
    - password_hash
    - phone
    type: object
  requests.CreateBasketRequest:
    properties:
      description:
        example: Ceci est un panier suprise
        type: string
      expiration_date:
        example: "2022-12-31"
        type: string
      name:
        example: panier surprise
        type: string
      original_price:
        example: 22
        type: number
      price:
        example: 8.99
        type: number
      quantity:
        example: 2
        type: integer
      restaurant_id:
        example: 1
        type: integer
      type:
        example: surprise
        type: string
    required:
    - name
    - original_price
    - price
    - quantity
    - restaurant_id
    - type
    type: object
...
  responses.BasketResponse:
    properties:
      address:
        type: string
      category:
        type: string
      discountPrice:
        type: number
      id:
        type: string
      name:
        type: string
      originalPrice:
        type: number
      rating:
        type: number
      typeBasket:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/baskets/:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all baskets
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.BasketResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all baskets
      tags:
      - Baskets
    post:
      consumes:
      - application/json
      description: Create a new basket with the provided details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Basket data
        in: body
        name: basket
        required: true
        schema:
          $ref: '#/definitions/requests.CreateBasketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad request, invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new basket
      tags:
      - Baskets
  /api/baskets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a basket by its ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Basket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid basket ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not authorized to delete this basket
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Basket not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a basket
      tags:
      - Baskets
    get:
      consumes:
      - application/json
      description: Retrieve a basket by its ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Basket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BasketResponse'
        "400":
          description: Invalid basket ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Basket not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a single basket
      tags:
      - Baskets
    put:
      consumes:
      - application/json
      description: Update a basket by its ID
      parameters:
      - description: Basket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Basket data
        in: body
        name: basket
        required: true
        schema:
          $ref: '#/definitions/models.Basket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Invalid basket ID or input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not authorized to update this basket
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Basket not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a basket
      tags:
      - Baskets
...
swagger: "2.0"